Pull reference tracking into its own utility class.
