+++ MWSE/TES3Object.cpp
@@ -38,6 +38,7 @@
 #include "TES3Quest.h"
 #include "TES3Race.h"
 #include "TES3Reference.h"
+#include "ReferenceTracker.h"
 #include "TES3Region.h"
 #include "TES3RepairTool.h"
 #include "TES3Script.h"
@@ -1005,294 +1006,8 @@ namespace TES3 {
 		return nullptr;
 	}
 
-	struct ReferenceTrackingData {
-		std::vector<Reference*> references = {};
-		bool dirtyLookup = false;
-	};
-
-	static std::unordered_set<Reference*> trackedReferences;
-	static std::unordered_map<const PhysicalObject*, ReferenceTrackingData> referenceDataByObject;
-	static std::unordered_map<const Cell*, size_t> referenceLookupCellOrder;
-	static const std::vector<Reference*> emptyReferences;
-	constexpr auto LOG_REFERENCE_TRACKING = false;
-
-	static const char* getReferenceLookupLogId(const BaseObject* object) {
-		__try {
-			return object ? object->getObjectID() : nullptr;
-		}
-		__except (EXCEPTION_EXECUTE_HANDLER) {
-			return "<invalid>";
-		}
-	}
-
-	static bool shouldLogReferenceLookupObject(const BaseObject* object) {
-		if constexpr (!LOG_REFERENCE_TRACKING) {
-			return false;
-		}
-
-		const auto id = getReferenceLookupLogId(object);
-		if (id == nullptr || id == reinterpret_cast<const char*>("<invalid>")) {
-			return false;
-		}
-
-		// Enter ID(s) here to track misbehavior.
-		return _stricmp(id, "TR_FM_Naureen") == 0;
-	}
-
-	static bool shouldLogReferenceLookupReference(const Reference* reference) {
-		if (reference == nullptr) {
-			return false;
-		}
-
-		return shouldLogReferenceLookupObject(reference->getBaseObject());
-	}
-
-	static const PhysicalObject* getLookupKey(const BaseObject* object) {
-		if (object == nullptr) {
-			return nullptr;
-		}
-
-		const auto baseObject = object->getBaseObject();
-		if (baseObject == nullptr) {
-			return nullptr;
-		}
-
-		return baseObject->asPhysicalObject();
-	}
-
-	static bool addReferenceToLookupImpl(Reference* reference) {
-		const auto key = getLookupKey(reference);
-		if (shouldLogReferenceLookupReference(reference)) {
-			mwse::log::getLog()
-				<< "[MWSE] Reference lookup index: ref=" << reference
-				<< " refId=" << (getReferenceLookupLogId(reference) ? getReferenceLookupLogId(reference) : "<no id>")
-				<< " base=" << reference->baseObject
-				<< " baseId=" << (getReferenceLookupLogId(reference->baseObject) ? getReferenceLookupLogId(reference->baseObject) : "<no id>")
-				<< " key=" << key
-				<< " keyId=" << (getReferenceLookupLogId(key) ? getReferenceLookupLogId(key) : "<no id>")
-				<< " cell=" << reference->getCell()
-				<< std::endl;
-		}
-
-		if (key == nullptr) {
-			if (shouldLogReferenceLookupReference(reference)) {
-				mwse::log::getLog() << "[MWSE] Reference lookup skipped: null key for ref=" << reference << std::endl;
-			}
-			return false;
-		}
-
-		auto& references = referenceDataByObject[key].references;
-		if (std::find(references.begin(), references.end(), reference) == references.end()) {
-			references.push_back(reference);
-			return true;
-		}
-
-		return false;
-	}
-
-	static void markReferenceLookupKeyDirty(const BaseObject* object) {
-		const auto key = getLookupKey(object);
-		if (key != nullptr) {
-			referenceDataByObject[key].dirtyLookup = true;
-		}
-	}
-
-	static void rebuildReferenceLookupCellOrder() {
-		referenceLookupCellOrder.clear();
-		const auto dataHandler = DataHandler::get();
-		const auto nonDynamicData = dataHandler ? dataHandler->nonDynamicData : nullptr;
-		const auto cells = nonDynamicData ? nonDynamicData->cells : nullptr;
-		if (cells == nullptr) {
-			return;
-		}
-
-		size_t index = 0;
-		for (const auto cell : *cells) {
-			if (cell != nullptr) {
-				referenceLookupCellOrder[cell] = index++;
-			}
-		}
-	}
-
-	static size_t getReferenceLookupCellOrder(const Cell* cell) {
-		if (cell == nullptr) {
-			return std::numeric_limits<size_t>::max();
-		}
-
-		auto orderIt = referenceLookupCellOrder.find(cell);
-		if (orderIt == referenceLookupCellOrder.end()) {
-			rebuildReferenceLookupCellOrder();
-			orderIt = referenceLookupCellOrder.find(cell);
-		}
-
-		return orderIt != referenceLookupCellOrder.end() ? orderIt->second : std::numeric_limits<size_t>::max();
-	}
-
-	static int getReferenceLookupListOrder(const Reference* reference) {
-		const auto cell = reference ? reference->getCell() : nullptr;
-		const auto list = reference ? reference->owningCollection.asReferenceList : nullptr;
-		if (cell == nullptr || list == nullptr) {
-			return 3;
-		}
-
-		if (list == &cell->actors) {
-			return 0;
-		}
-		if (list == &cell->persistentRefs) {
-			return 1;
-		}
-		if (list == &cell->temporaryRefs) {
-			return 2;
-		}
-		return 3;
-	}
-
-	static bool isReferenceBeforeInCurrentList(const Reference* lhs, const Reference* rhs) {
-		const auto list = lhs ? lhs->owningCollection.asReferenceList : nullptr;
-		if (list == nullptr || list != (rhs ? rhs->owningCollection.asReferenceList : nullptr)) {
-			return lhs < rhs;
-		}
-
-		for (const auto reference : *list) {
-			if (reference == lhs) {
-				return true;
-			}
-			if (reference == rhs) {
-				return false;
-			}
-		}
-
-		return lhs < rhs;
-	}
-
-	static bool isReferenceBeforeInVanillaLookupOrder(const Reference* lhs, const Reference* rhs) {
-		const auto lhsCellOrder = getReferenceLookupCellOrder(lhs ? lhs->getCell() : nullptr);
-		const auto rhsCellOrder = getReferenceLookupCellOrder(rhs ? rhs->getCell() : nullptr);
-		if (lhsCellOrder != rhsCellOrder) {
-			return lhsCellOrder < rhsCellOrder;
-		}
-
-		const auto lhsListOrder = getReferenceLookupListOrder(lhs);
-		const auto rhsListOrder = getReferenceLookupListOrder(rhs);
-		if (lhsListOrder != rhsListOrder) {
-			return lhsListOrder < rhsListOrder;
-		}
-
-		return isReferenceBeforeInCurrentList(lhs, rhs);
-	}
-
-	static void sortReferencesLookupForKey(const PhysicalObject* key) {
-		if (key == nullptr) {
-			return;
-		}
-
-		const auto dataIt = referenceDataByObject.find(key);
-		if (dataIt == referenceDataByObject.end() || !dataIt->second.dirtyLookup) {
-			return;
-		}
-
-		if constexpr (LOG_REFERENCE_TRACKING) {
-			mwse::log::getLog()
-				<< "[MWSE] Sorting reference lookup: key=" << key
-				<< " keyId=" << (getReferenceLookupLogId(key) ? getReferenceLookupLogId(key) : "<no id>")
-				<< std::endl;
-		}
-
-		auto& references = dataIt->second.references;
-		std::sort(references.begin(), references.end(), isReferenceBeforeInVanillaLookupOrder);
-		dataIt->second.dirtyLookup = false;
-	}
-
 	const std::vector<Reference*>& PhysicalObject::getReferences() const {
-		const auto key = getLookupKey(this);
-		sortReferencesLookupForKey(key);
-
-		const auto logLookup = shouldLogReferenceLookupObject(this) || shouldLogReferenceLookupObject(key);
-		if (logLookup) {
-			mwse::log::getLog()
-				<< "[MWSE] Reference lookup query: object=" << this
-				<< " objectId=" << (getReferenceLookupLogId(this) ? getReferenceLookupLogId(this) : "<no id>")
-				<< " key=" << key
-				<< " keyId=" << (getReferenceLookupLogId(key) ? getReferenceLookupLogId(key) : "<no id>")
-				<< std::endl;
-		}
-
-		if (key == nullptr) {
-			if (logLookup) {
-				mwse::log::getLog() << "[MWSE] Reference lookup query result: null key" << std::endl;
-			}
-			return emptyReferences;
-		}
-
-		const auto referencesIt = referenceDataByObject.find(key);
-		if (referencesIt == referenceDataByObject.end()) {
-			if (logLookup) {
-				mwse::log::getLog() << "[MWSE] Reference lookup query result: missing key" << std::endl;
-			}
-			return emptyReferences;
-		}
-
-		if (logLookup) {
-			mwse::log::getLog() << "[MWSE] Reference lookup query result: count=" << referencesIt->second.references.size() << std::endl;
-		}
-
-		return referencesIt->second.references;
-	}
-
-	void PhysicalObject::trackReferenceForLookup(Reference* reference) {
-		if (reference == nullptr) {
-			return;
-		}
-
-		if (trackedReferences.insert(reference).second) {
-			addReferenceToLookupImpl(reference);
-			markReferenceLookupKeyDirty(reference);
-			if (shouldLogReferenceLookupReference(reference)) {
-				mwse::log::getLog()
-					<< "[MWSE] Reference lookup tracked: ref=" << reference
-					<< " refId=" << (getReferenceLookupLogId(reference) ? getReferenceLookupLogId(reference) : "<no id>")
-					<< " base=" << reference->baseObject
-					<< " baseId=" << (getReferenceLookupLogId(reference->baseObject) ? getReferenceLookupLogId(reference->baseObject) : "<no id>")
-					<< " trackedReferences=" << trackedReferences.size()
-					<< std::endl;
-			}
-		}
-	}
-
-	void PhysicalObject::untrackReferenceForLookup(Reference* reference) {
-		if (reference == nullptr) {
-			return;
-		}
-
-		const auto key = getLookupKey(reference);
-		markReferenceLookupKeyDirty(reference);
-		if (trackedReferences.erase(reference)) {
-			if (key != nullptr) {
-				const auto referencesIt = referenceDataByObject.find(key);
-				if (referencesIt != referenceDataByObject.end()) {
-					auto& references = referencesIt->second.references;
-					references.erase(std::remove(references.begin(), references.end(), reference), references.end());
-				}
-			}
-
-			if (shouldLogReferenceLookupReference(reference)) {
-				mwse::log::getLog()
-					<< "[MWSE] Reference lookup untracked: ref=" << reference
-					<< " refId=" << (getReferenceLookupLogId(reference) ? getReferenceLookupLogId(reference) : "<no id>")
-					<< " base=" << reference->baseObject
-					<< " baseId=" << (getReferenceLookupLogId(reference->baseObject) ? getReferenceLookupLogId(reference->baseObject) : "<no id>")
-					<< " trackedReferences=" << trackedReferences.size()
-					<< std::endl;
-			}
-		}
-	}
-
-	void PhysicalObject::markReferencesLookupDirty(const BaseObject* object) {
-		if (object == nullptr) {
-			return;
-		}
-
-		markReferenceLookupKeyDirty(object);
+		return ReferenceTracker::getReferences(this);
 	}
 }
